[#3958] Improve localization of rest messages
Updates the rest message to select the proper plural form of Hit
Points and Hit Dice to match the number recovered.

Also adjusts the flavor to automatically display the rest time in
the message flavor, rather than relying on hard-coded labels for
each rest variant. This ensures the flavor will properly display
the rest time even if it doesn't exactly match the default values.

In order to support this as well as improve other time-based
localization in the future, this PR adds two new utlity methods:
- `convertTime` converts a time from one set of units to another.
  It will also attempt to pick the best unit automatically if the
  `to` units are not defined, selecting the largest unit that can
  fully represent the value (with some wiggle room so it will
  display 90 minutes, rather than 1.5 hours).
- `formatTime` converts a time value and units into a string for
  display, relying on the system-provided plural localization if
  available and if not using Javascript's standard unit formatting
  for numbers.