[#4708] Convert `AdvancementManager` to V2 app & styles
Converts `AdvancementManager` to use `ApplicationV2` and V2 system
styles. Because old V1 advancement flow applications are embedded
within this application, there are some V1 classes added to ensure
that styling of old flows aren't completely broken.

This also takes a number of soft-private properties and methods on
the manager and makes them hard-private.