[#3357] Support variable types & multiple damage parts in enricher
Adds support for variable damage types when rolling damage through
a damage enricher as well as multiple damage parts.

To indicate variable damage types, the explicit `type` key must be
used when writing the enricher and the different types must be
separated by a pipe (`|`) character:

```
[[/damage 2d6 type=fire|cold]]
```

To specify multiple damage parts the enricher config is split by
an amperand (`&`), with each side supporting its own formula and
damage types:

```
[[/damage 1d6 fire & 1d6 cold]]
[[/damage 1d6 fire & 1d6 cold average]]
```

Any configuration options that apply to the whole enricher can be
placed anywhere in the configuration.

If the average option is used and there are exactly two damage
parts, then they will be written as "A plus B", otherwise they
will be written "A and B". This is to match the format of NPC
statblocks.

Closes #3357